Change parser to strip DOS newlines from comments
CRLF line endings were causing comments to have trailing `\r`
characters, this fix implements a check within the comment parsing
algorithms to determine if CRLF line endings were used. If so, `pos` is
decremented to account for the extra character.

New tests were added with the `sass` file saved as a DOS file with CRLF
line endings. Just as with the other CRLF test files, changing the line
endings automatically via an IDE or similar will cause the test to fail.

Fixes #224